feat(core): promote Stepper from Lab
