+++ .changeset/promote-stepper-to-core.md
@@ -1,0 +1,8 @@
+---
+'@astryxdesign/cli': patch
+'@astryxdesign/core': patch
+---
+
+[component] Promote `Stepper` and `Step` from the canary-only Lab package to Core. The stable package now ships their existing horizontal/vertical layouts, separated and on-track indicators, semantic status, density, non-linear navigation, and narrow-width label collapsing, plus Core documentation and rendered examples. The default `aria-label` is now localized.
+
+@ernestt
