--- packages/lab/src/Stepper/Step.doc.mjs
@@ -1,17 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('@astryxdesign/cli/authoring').ComponentDoc} */
-
-export const docs = {
-  name: 'Step',
-  displayName: 'Step',
-  group: 'Stepper',
-  category: 'Navigation',
-  isHiddenFromOverview: true,
-
-  usage: {
-    description: '',
-  },
-
-  props: [],
-};
