+++ packages/lab/src/index.ts
@@ -240,7 +240,6 @@ export {
 
 // Chart v2 (config model) moved to its own package: @astryxdesign/charts.
 // It is no longer re-exported from @astryxdesign/lab.
-export * from './Stepper';
 export * from './CircularProgress';
 
 // ListInput — compact editor for short collections of simple records
